Revert back to extending parent azure step so can share undo step
